feat: moved all the api calls to services
