+++ src/services/index.js
@@ -1,0 +1,2 @@
+export * from './cartServices'
+export * from './productServices'
